[feature/mainpage] feat : useAuth() hook의 user 정보 호출
